Disable proto rule generation for bleve
The bleve project generates their own golang code from their proto
definition and doesn't need Bazel to generate it. This change disables
proto rule generation for the bleve dependency so that the .pb.go file
gets added to the sources for the upsidedown library within bleve. This
allows the project to build successfully.